+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Rename identifiers
 - Give each dot file a unique name
 - Highlight selected node in Node.writeImageFile, childScopeDepth is -1 by default
+- Instantiation of a scope requires a blue print
 
 ### Fixed